Salt based docker deployment
This commit adds a first attempt at a deployment system for containers
using salt. It adds an example pillar file that shows how to define
the nginx vhosts and container settings. It adds a state which defines
an upstart job to run the container. It adds a state which creates an
environment file for each container based on pillar/grain data.
It adds a simple nginx proxy which proxies to the containers and
S3 for assets.