Fix for AWS throttling.
AWS throttles requestst to its API. This can cause the queries in the
custom grains to fail. If they fail, then the grain targeting in the
pillar fails and we can get unexpected behaviour.

This puts a retry loop with some random wait if we get throttled and
also adds a check to the pillar top files which breaks the pillar if
the custom grain failed.