Fix a bash injection exploit.
We were allowing users to set a grain, then including this grain in an
upstart job. This allowed the user to set the grain to include a bash
string (e.g. grains.setval my_task '; cat /etc/passwd' and become root.
This fix puts single quotes in the jinja template and pipes the grain
throgh a yaml filter (preventing the user from closing the single
quotes.