Prevent arbitrary arguments on docker run
Single quotes prevent bash injection using the task or version grain but
they still get passed to docker run. This means a user can run their
container and mount a file from the host etc. To prevent this we change
here to move the task to the environment file and to pass the version to
docker pull and docker tag (which are less dangerous than docker run).