Allow non-nginx enabled containers
All containers are proxied by nginx automatically, this is not always
what we want. This change allows containers to be specified outside
the vhost section in the pillar so that they are set up independently
from nginx. Also, location and ports container pillar settings are
made optional, if we don't specify them then we don't do the related
nginx setup. (Closes ministryofjustice/template-deploy#70)

- Separate out proxied and non-proxied containers
- Only configure nginx for proxied containers
- Abstract out container generation jinja into macros
- Updated pillar documentaion
- Link template and pillar doc example code directly to code templates
- Use redis container for non-proxied container
- Abstract ELB registration code for containers into macro
- Abstract container environment variable setup into macro