Make branch runner restart changed branches
This switches docker run to an upstart job. This is so that when a
branch is pushed, running the branch runner will pull the container and
restart it. Because we have to override some env vars in the branch
containers and to keep the jinja sane(ish) i've split the branch
containers upstart template from the regular template.