Pass full container name to docker pull
The container dictionary has a key (short name) which we use for
creating files (as it doesn't contain a slash). It also has a name
property which can include a slash (we pass this one to docker
pull/run).

This fix ensures that we pass the correct long name to the docker pull
and run used by the branch runner.