Tag attached EBS volumes
AutoScalingGroups tag EC2 instances but not volumes. Since salt will
highstate when a machine boots then we can do the tagging here.

In order to succeed, this operation requires the ec2:CreateTags perm
from the instance itself. In the event of this permission being missing
salt will continue, obliviously.

Useful though they are, tags are used only for auditing purposes. They
do not affect operation of a stack.