Set a default nginx host to avoid branch confusion
When using branchrunners, nginx treats the first defined `server` (so
usually the first file in `/etc/nginx/conf.d/`) listening to a given
port as the default host if no `server_name` match is found. This means
that a branch called `aaa` (or anything alphabetically before the
`server_name`) might override the main app host.

Instead, we want the default host to be explicitly set to a
non-branchbuilder host.