Add a fallback check to the pillar for a containers host port
Currently we use the docker.port call for salt to get a containers
actual exposed host port. When we have called --net="host" with
docker though, this will be empty, and the call will fail, even
though the containers port is still exposed in the instance. This
change will add a fallback call to the pillar, so that we can catch
this situation and try our best to get the host port if possible.