Allow container linking
Manually passing linking options to docker as args in the pillar will
set up linked containers sometimes. But since containers need to be up
before they are linked to, this requires a start order to run containers
in. Since we use upstart jobs, we can enable linking in the pillar and
alter the jobs to take this order into account, requiring the necessary
linked to containers to be running before the service will successfully
start, and reacting to those services restarting to restart and therefore
recover linking to the restarted container.

* Add links pillar option
* links_restarts sets upstart start stanza to rely on linked container jobs
* Set upstart container job to link to containers specified in pillar
* Create pre-start upstart section to handle dependent job status
* Alter post-start script to use variable for max-retries for loop