Change container linking pillar structure
The current structure is a little unclear, this change will rewrite the
required pillar structure in the format.

links:
  - link: sidekiq
    required: True

This will also alter the behaviour of the container service, by default
'required' will be true, meaning that the linking container will not
be started unless all of its link target containers are available.

If 'required' is set to False this means that we are defining the target
container link as non-essential to the running of the linking container.
This is unlikely to be the behaviour desired in most cases, but may have
relevance where the linking container runs a service that will become
degraded when the linked-to container is not available, but it is still
preferable to run a degraded service than run no service at all.